ABBCK-38 # Poprawa widoku magazyniera, dodanie wyświetlania zawartości tabel

+++ src/main/java/hibernate/entity/Pracownicy.java
@@ -36,7 +36,7 @@ public class Pracownicy {
     public Pracownicy() {
     }
 
-    public long getIdPracownika() {
+    public int getIdPracownika() {
         return idPracownika;
     }
